server/requirements.txt
a==1.0
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
astor==0.8.1
astunparse==1.6.3
attrs==25.3.0
backup==0.0.1
BareNecessities==0.2.8
bcrypt==4.3.0
beautifulsoup4==4.13.3
bidict==0.23.1
black==23.7.0
blinker==1.9.0
blis==1.2.0
Brotli==1.1.0
build==1.3.0
camera==1.3.0
cartes==0.8.4
Cartopy==0.24.1
catalogue==2.0.10
catboost==1.2.8
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
cmaes==0.12.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
config==0.5.1
contourpy==1.3.1
cors==1.0.1
crypto==1.4.1
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
dashboard==0.0.6
data_preprocessing==0.0.2
databases==0.9.0
DateTime==5.5
db==0.1.1
dearpygui==2.0.0
distlib==0.3.9
Django==5.1.7
environment==1.0.0
et_xmlfile==2.0.0
fastapi==0.115.11
filelock==3.18.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
flexcache==0.3
flexparser==0.4
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2025.3.0
funcs==0.11.0
future==1.0.0
gast==0.6.0
gaussian==0.1
geopandas==1.0.1
gevent==24.11.1
google==3.0.0
google-pasta==0.2.0
graphviz==0.21
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
image==1.5.33
impunity==1.0.5
imutils==0.5.4
inflate64==1.0.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.20.0
jsonschema-specifications==2024.10.1
keras==3.9.0
keras-models==0.0.7
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.45.0
lstm==0.1.0
lxml==5.3.1
lz4==4.4.3
Mail==2.1.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mashumaro==3.15
matplotlib==3.10.0
mdurl==0.1.2
metar==1.11.0
minio==7.2.15
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
multivolumefile==0.2.3
murmurhash==1.0.12
mypy_extensions==1.1.0
Naked==0.1.32
named==1.4.2
namex==0.0.8
narwhals==1.31.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.62.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.0
openap==2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
optuna==3.3.0
orjson==3.10.15
orm==0.2.0.dev1
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pathspec==0.12.1
pillow==11.1.0
Pint==0.24.4
pitot==0.3.2
platformdirs==4.3.6
plotly==6.3.0
pluggy==1.5.0
preshed==3.0.9
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.9
py7zr==0.22.0
pyarrow==19.0.1
pybcj==1.0.3
pycparser==2.22
pycryptodome==3.22.0
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.8.0
pyogrio==0.10.0
pyopensky==2.14
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyppmd==1.1.1
pyproj==3.7.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyshp==2.3.1
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.1
pyzstd==0.16.2
rectangle==0.5.1
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.23.1
rs1090==0.4.7
schedule==1.2.2
schema==0.7.7
scikit-learn==1.6.1
scipy==1.11.2
seaborn==0.13.2
sequential==1.0.0
setuptools==75.8.2
shap==0.48.0
shapely==2.0.7
shellescape==3.8.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
slicer==0.0.8
smart-open==7.1.0
sniffio==1.3.1
solus==1.2.2
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.39
sqlparse==0.5.3
srsly==2.5.1
starlette==0.46.1
sumolib==1.22.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow_intel==2.18.0
termcolor==2.5.0
texttable==1.7.0
the==0.1.5
thinc==8.3.4
thread==2.0.5
threadpoolctl==3.5.0
timedelta==2020.12.3
timestamp==0.0.1
tldextract==5.1.3
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.66.1
traci==1.22.0
traffic==2.12
trafficSimulator==0.0.1
trino==0.333.0
typer==0.15.2
typesystem==0.3.0.dev0
typing-aliases==1.10.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
vehicle==2.2.2
virtualenv==20.30.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wraps==0.15.0
wrapt==1.17.2
wsproto==1.2.0
xgboost==3.0.5
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
